Refactor wallet persister interfaces
