+++ lib/wasm/src/wallet_persister/node_fs.rs
@@ -1,0 +1,1 @@
+
